Part7-8

+++ test_db.py
@@ -1,8 +1,8 @@
 import subprocess
 import os
 
-def run_script(commands):
-    if os.path.exists('test.db'):
+def run_script(commands, clear_db = True):
+    if os.path.exists('test.db') and clear_db:
         os.remove('test.db')
 
     process = subprocess.Popen(
@@ -99,7 +99,7 @@ def test_keeps_data_after_closing_connection():
     result2 = run_script([
         "select",
         ".exit",
-    ])
+    ], False)
 
     expected2 = [
         "db > (1, user1, person1@example.com)",
@@ -108,3 +108,41 @@ def test_keeps_data_after_closing_connection():
     ]
 
     assert expected2 == result2
+
+def test_prints_constants():
+    result = run_script([
+        ".constants",
+        ".exit",
+    ])
+
+    expected = [
+        "db > Constants:",
+        "ROW_SIZE: 293",
+        "COMMON_NODE_HEADER_SIZE: 6",
+        "LEAF_NODE_HEADER_SIZE: 10",
+        "LEAF_NODE_CELL_SIZE: 297",
+        "LEAF_NODE_SPACE_FOR_CELLS: 4086",
+        "LEAF_NODE_MAX_CELLS: 13",
+        "db > "
+        ]
+    
+    assert expected == result
+
+
+def test_allows_printing_out_the_structure_of_a_one_node_btree():
+    script = [f"insert {i} user{i} person{i}@example.com" for i in [3,1,2]]
+    script.extend([".btree", ".exit"])
+    result = run_script(script)
+    expected = [
+        "db > Executed.",
+        "db > Executed.",
+        "db > Executed.",
+        "db > Tree:",
+        "Leaf (size: 3)",
+        "  - 0 : 3",
+        "  - 1 : 1",
+        "  - 2 : 2",
+        "db > "
+        ]
+    
+    assert expected == result
